+++ test/easy_qr_code_test.dart
@@ -1,21 +1,22 @@
 import 'dart:typed_data';
-import 'dart:ui';
+import 'dart:ui' as ui;
+import 'package:easy_qr_code/src/models/qr_painter_with_background.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:easy_qr_code/easy_qr_code.dart'; // Ensure this is the correct path
-import 'package:qr_flutter/qr_flutter.dart';
-
+import 'package:easy_qr_code/easy_qr_code.dart';
 
 void main() {
-  group('QRGenerator Tests', () {
-    test('should generate a QR code widget successfully', () async {
+  group('EasyQRCodeGenerator Tests', () {
+    testWidgets('should generate a QR code widget successfully',
+        (WidgetTester tester) async {
       // Arrange
       final qrGenerator = EasyQRCodeGenerator();
       const testData = 'https://example.com';
       const expectedSize = 200.0;
 
       // Act
-      final qrWidget = await qrGenerator.generateQRCodeWidget(data: testData, size: expectedSize);
+      final qrWidget = await qrGenerator.generateQRCodeWidget(
+          data: testData, size: expectedSize);
 
       // Assert
       expect(qrWidget, isA<SizedBox>());
@@ -23,56 +24,71 @@ void main() {
       expect(sizedBox.width, expectedSize);
       expect(sizedBox.height, expectedSize);
 
-      // You can also check for a CustomPaint widget if needed:
+      // Verify the child is a CustomPaint widget
+      expect(sizedBox.child, isA<CustomPaint>());
       final customPaint = sizedBox.child as CustomPaint;
-      expect(customPaint.painter, isA<QrPainter>());
+      expect(customPaint.painter, isA<QrPainterWithBackground>());
+    });
+
+    test('should generate a QR code image successfully', () async {
+      // Arrange
+      final qrGenerator = EasyQRCodeGenerator();
+      const testData = 'https://example.com';
+      const expectedSize = 200.0;
+
+      // Act
+      final qrImage = await qrGenerator.generateQRCodeImage(
+          data: testData, size: expectedSize);
+
+      // Assert
+      expect(qrImage, isA<ui.Image>());
+      expect(qrImage.width, expectedSize.toInt());
+      expect(qrImage.height, expectedSize.toInt());
     });
   });
 
   group('EasyQRCodeReader Tests', () {
     test('should decode a valid QR code from image bytes', () async {
-      // Arrange: Generate a simple image from the string 'https://example.com'
+      // Arrange
       const testData = 'https://example.com';
       final qrGenerator = EasyQRCodeGenerator();
 
-      // Generate the QR code widget (this will be used to get the image bytes)
-      final qrWidget = await qrGenerator.generateQRCodeWidget(data: testData);
-      final sizedBox = qrWidget as SizedBox;
-      final customPaint = sizedBox.child as CustomPaint;
-      final painter = customPaint.painter as QrPainter;
+      // Generate the QR code image
+      final qrImage = await qrGenerator.generateQRCodeImage(data: testData);
+      final imageBytes = await _getImageBytesFromImage(qrImage);
 
-      // Convert the QR code to a byte array for testing the decoder
-      final imageBytes = await _getImageBytes(painter);
+      // Verify bytes are not empty
+      expect(imageBytes.isNotEmpty, isTrue,
+          reason: 'Image bytes should not be empty.');
+
       final qrReader = EasyQRCodeReader();
 
-      // Act: Try decoding the QR code
+      // Act
       final decodedText = await qrReader.decode(Uint8List.fromList(imageBytes));
 
-      // Assert: The decoded text should match the input string
+      // Assert
       expect(decodedText, testData);
     });
 
     test('should return null when the QR code image is invalid', () async {
-      // Arrange: Provide invalid bytes (empty data)
+      // Arrange
       final qrReader = EasyQRCodeReader();
       final invalidBytes = Uint8List(0);
 
-      // Act: Attempt to decode the invalid QR code
+      // Act
       final decodedText = await qrReader.decode(invalidBytes);
 
-      // Assert: The result should be null
+      // Assert
       expect(decodedText, isNull);
     });
   });
 }
 
-/// Helper method to get image bytes from a QrPainter
-Future<List<int>> _getImageBytes(QrPainter painter) async {
-  final recorder = PictureRecorder();
-  final canvas = Canvas(recorder);
-  painter.paint(canvas, const Size(200, 200)); // Assuming a 200x200 size for the QR code
-  final picture = recorder.endRecording();
-  final img = await picture.toImage(200, 200); // Same size as above
-  final byteData = await img.toByteData(format: ImageByteFormat.png);
-  return byteData!.buffer.asUint8List();
+/// Helper method to convert a ui.Image to a byte array for testing
+Future<List<int>> _getImageBytesFromImage(ui.Image image) async {
+  final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
+  if (byteData == null || byteData.lengthInBytes == 0) {
+    throw Exception('Failed to convert ui.Image to byte data.');
+  }
+  return byteData.buffer.asUint8List();
 }