+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+## 0.1.0
+* Added QR code generation with an image display (both widget and image formats).
+* Implemented the ability to save generated QR code images to the device.
+* Added the option to share the generated QR code image.
 
-* TODO: Describe initial release.
+
+## 0.0.1
+* initial release.
+* Enabled the ability to read QR codes from the gallery and camera.
+* Improved UI with clearer layout and more intuitive flow for scanning and generating QR codes.